pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.5",
    "google-search-results>=2.4.2",
    "langchain>=1.2.9",
    "langchain-openai>=1.1.7",
    "pinecone>=8.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "serpapi>=0.1.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
